+++ KollektivSystem.ApiService/Models/Dtos/TicketTypes/TicketTypeResponse.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; init; }
         public string Name { get; init; } = null!;
-        public int Price { get; init; }
+        public float Price { get; init; }
         public TimeSpan AliveTime { get; init; }
         public bool IsActive { get; init; }
     }
